requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0

regex>=2024.11.0
pytest>=8.3.0

# LLM Provider SDKs
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0

